Fixed large data sending problem. See details
I'm not very happy with this fix. The problem seems to go away if the
file data is read into a buffer and then written to the connection with
io.Copy, not io.CopyN. I don't know why this fixes it, or even where
the problem is.

As a temporary measure, I'm pushing this fix, but I'd like to identify
the problem and make a better fix.