Breaking Changes. See details.
Substantial changes, removing spdy.Request and spdy.Header, replacing
them with their net/http equivalents. This required refactoring of much
of the code, but the result is far simpler and makes better use of
existing code.

The breaking change is that Handlers now receive an http.Request, not a
spdy.Request.