Updated docs
fixes #60. CHANGELOG: The structure of the package has changed significantly. Multiple sub-packages have been made to help distribute the code and improve structure. This has generally left the core public API unaffected.
The main impact that might affect users is that most of the public interfaces have been moved to spdy/common. The protocol-specific code has been moved to separate packages; namely spdy/spdy2 and spdy/spdy3. This means that more of the code is exported overall, but the base spdy package is much cleaner.