Fixed bug in connection flow control.
Fixes #82.

This was a silly bug. The connection-level flow control was updating
the window correctly but was buffering the whole frame, so the peer
had no idea we were constrained if it was caused by one big frame.
Now we handle it as with stream flow control, where we send as much
as we can and buffer what's left.

Thanks to Rob Napier for triggering the bug; it's pretty embarrassing
that it was there, but that's what you get. The regression test
*should* make sure nothing similar crops up later.