Fixed ReadFrom/WriteTo returns
As found by jeffyang28, there were some issues in the returns
from some of the frames' ReadFrom and WriteTo methods. Rather than
continuing to rely on maths done by hand, I've switched to a more
idiomatic solution.

In common I've added ReadCounter and WriteCounter, which are a pair
of helper structures which just keep track of the number of bytes
read from or written to an io.Reader/Writer.

I've used these in all of the frame I/O methods to ensure that the
value returned is the actual number of bytes read/written.

Fixes #94